refactor(groove): split database responsibilities
